+++ metadata.rb
@@ -2,9 +2,9 @@ name             'loggly_rsyslog_ng'
 maintainer       'Kostiantyn Lysenko'
 maintainer_email 'gshaud@gmail.com'
 license          'Apache 2.0'
-description      'Configures rsyslog to send logs to Loggly'
+description      'LWRP for configuring rsyslog to send data to Loggly'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '3.0.0'
+version          '4.0.0'
 
 supports 'debian'